GoogTalkFunkBotBQ/requirements.txt
python-telegram-bot==13.4.1
pandas
pyarrow
sklearn
google
pandas_gbq
